requirements.txt
selenium>=4.6.0
pytest>=7.0.0
allure-pytest>=2.13.0
